==> 1.1_corpus_creation.Rmd <==
# creation of the corpus

### similarities, VSM, CLUSERING

j'ai créé 2 corpus
- avant / après
- un seul avec tous les tweets
```{r}
# trump analysis
TRUMP_all_tweet <- read_csv("data supervised_L/trump_after.csv")

#get only the tweets after the election
Trump_after_analysis <- TRUMP_all_tweet %>% 
  mutate(date = date(created_at)) %>% filter(date >= "2020-11-03", 
                                             display_text_width > 100)
#get the tweet before
Trump_before_analysis <- TRUMP_all_tweet %>% 
  mutate(date = date(created_at)) %>% filter(date < "2020-11-03", 
                                             display_text_width > 100)
#add variable before or after
Trump_after_analysis <-Trump_after_analysis %>% 
  mutate(when = "after")
Trump_before_analysis <-Trump_before_analysis %>% 
  mutate(when = "before")
#############

#book of tweets "after" 
Trump_after_analysis <- Trump_after_analysis %>% 
  select(text,screen_name, when) %>% 
  group_by(screen_name, when) %>% 
  summarise(paste(Trump_after_analysis$text, collapse = '. ')) 

Trump_after_analysis <- rename(Trump_after_analysis, text = `paste(Trump_after_analysis$text, collapse = ". ")`)
#book of tweets "before"  
Trump_before_analysis <- Trump_before_analysis %>% 
  select(text,screen_name, when) %>% 
  group_by(screen_name, when) %>% 
  summarise(paste(Trump_before_analysis$text, collapse = '. ')) 

Trump_before_analysis <- rename(Trump_before_analysis, text = `paste(Trump_before_analysis$text, collapse = ". ")`)


# vp analysis
vp_all_tweet <- read_csv("data supervised_L/vp_after.csv")
#get only the tweets after the election
vp_after_analysis <- vp_all_tweet %>% 
  mutate(date = date(created_at)) %>% 
  filter(date >= "2020-11-03",
        display_text_width > 100)
#get the tweet before
vp_before_analysis <- vp_all_tweet %>% 
  mutate(date = date(created_at)) %>% 
  filter(date < "2020-11-03",
        display_text_width > 100)
#add variable before or after
vp_after_analysis <-vp_after_analysis %>% 
  mutate(when = "after")

vp_before_analysis <-vp_before_analysis %>% 
  mutate(when = "before")
#############
vp_after_analysis <- vp_after_analysis %>% 
  select(text,screen_name, when) %>% 
  group_by(screen_name, when) %>% 
  summarise(paste(vp_after_analysis$text, collapse = '. ')) 

vp_after_analysis <- rename(vp_after_analysis, text = `paste(vp_after_analysis$text, collapse = ". ")`)
  
vp_before_analysis <- vp_before_analysis %>% 
  select(text,screen_name, when) %>% 
  group_by(screen_name, when) %>% 
  summarise(paste(vp_before_analysis$text, collapse = '. ')) 

vp_before_analysis <- rename(vp_before_analysis, text = `paste(vp_before_analysis$text, collapse = ". ")`)

# biden analysis
biden_all_tweet <- read_csv("data supervised_L/biden_after.csv")
#get only the tweets after the election
biden_after_analysis <- biden_all_tweet %>% 
  mutate(date = date(created_at)) %>% 
  filter(date >= "2020-11-03",
        display_text_width > 100)
#get the tweet before
biden_before_analysis <- biden_all_tweet %>% 
  mutate(date = date(created_at)) %>% 
  filter(date < "2020-11-03",
        display_text_width > 100)
#add variable before or after
biden_after_analysis <-biden_after_analysis %>% 
  mutate(when = "after")

biden_before_analysis <-biden_before_analysis %>% 
  mutate(when = "before")
#############
biden_after_analysis <- biden_after_analysis %>% 
  select(text,screen_name, when) %>% 
  group_by(screen_name, when) %>% 
  summarise(paste(biden_after_analysis$text, collapse = '. ')) 

biden_after_analysis <- rename(biden_after_analysis, text = `paste(biden_after_analysis$text, collapse = ". ")`)
  
biden_before_analysis <- biden_before_analysis %>% 
  select(text,screen_name, when) %>% 
  group_by(screen_name, when) %>% 
  summarise(paste(biden_before_analysis$text, collapse = '. ')) 

biden_before_analysis <- rename(biden_before_analysis, text = `paste(biden_before_analysis$text, collapse = ". ")`)

# kamala analysis
kamala_all_tweet <- read_csv("data supervised_L/kamala_after.csv")
#get only the tweets after the election
kamala_after_analysis <- kamala_all_tweet %>% 
  mutate(date = date(created_at)) %>% 
  filter(date >= "2020-11-03",
        display_text_width > 100)
#get the tweet before
kamala_before_analysis <- kamala_all_tweet %>% 
  mutate(date = date(created_at)) %>% 
  filter(date < "2020-11-03",
        display_text_width > 100)
#add variable before or after
kamala_after_analysis <-kamala_after_analysis %>% 
  mutate(when = "after")

kamala_before_analysis <-kamala_before_analysis %>% 
  mutate(when = "before")
#############
kamala_after_analysis <- kamala_after_analysis %>% 
  select(text,screen_name, when) %>% 
  group_by(screen_name, when) %>% 
  summarise(paste(kamala_after_analysis$text, collapse = '. ')) 

kamala_after_analysis <- rename(kamala_after_analysis, text = `paste(kamala_after_analysis$text, collapse = ". ")`)
  
kamala_before_analysis <- kamala_before_analysis %>% 
  select(text,screen_name, when) %>% 
  group_by(screen_name, when) %>% 
  summarise(paste(kamala_before_analysis$text, collapse = '. ')) 

kamala_before_analysis <- rename(kamala_before_analysis, text = `paste(kamala_before_analysis$text, collapse = ". ")`)

# obama analysis
obama_all_tweet <- read_csv("data supervised_L/obama_after.csv")
#get only the tweets after the election
obama_after_analysis <- obama_all_tweet %>% 
  mutate(date = date(created_at)) %>% 
  filter(date >= "2020-11-03",
        display_text_width > 100)
#get the tweet before
obama_before_analysis <- obama_all_tweet %>% 
  mutate(date = date(created_at)) %>% 
  filter(date < "2020-11-03",
        display_text_width > 100)
#add variable before or after
obama_after_analysis <-obama_after_analysis %>% 
  mutate(when = "after")

obama_before_analysis <-obama_before_analysis %>% 
  mutate(when = "before")
#############
obama_after_analysis <- obama_after_analysis %>% 
  select(text,screen_name, when) %>% 
  group_by(screen_name, when) %>% 
  summarise(paste(obama_after_analysis$text, collapse = '. ')) 

obama_after_analysis <- rename(obama_after_analysis, text = `paste(obama_after_analysis$text, collapse = ". ")`)
  
obama_before_analysis <- obama_before_analysis %>% 
  select(text,screen_name, when) %>% 
  group_by(screen_name, when) %>% 
  summarise(paste(obama_before_analysis$text, collapse = '. ')) 

obama_before_analysis <- rename(obama_before_analysis, text = `paste(obama_before_analysis$text, collapse = ". ")`)


# AOC analysis
aoc_all_tweet <- read_csv("data supervised_L/aoc_after.csv")
#get only the tweets after the election
aoc_after_analysis <- aoc_all_tweet %>% 
  mutate(date = date(created_at)) %>% 
  filter(date >= "2020-11-03",
        display_text_width > 100)
#get the tweet before
aoc_before_analysis <- aoc_all_tweet %>% 
  mutate(date = date(created_at)) %>% 
  filter(date < "2020-11-03",
        display_text_width > 100)
#add variable before or after
aoc_after_analysis <-aoc_after_analysis %>% 
  mutate(when = "after")

aoc_before_analysis <-aoc_before_analysis %>% 
  mutate(when = "before")
#############
aoc_after_analysis <- aoc_after_analysis %>% 
  select(text,screen_name, when) %>% 
  group_by(screen_name, when) %>% 
  summarise(paste(aoc_after_analysis$text, collapse = '. ')) 

aoc_after_analysis <- rename(aoc_after_analysis, text = `paste(aoc_after_analysis$text, collapse = ". ")`)
  
aoc_before_analysis <- aoc_before_analysis %>% 
  select(text,screen_name, when) %>% 
  group_by(screen_name, when) %>% 
  summarise(paste(aoc_before_analysis$text, collapse = '. ')) 

aoc_before_analysis <- rename(aoc_before_analysis, text = `paste(aoc_before_analysis$text, collapse = ". ")`)

corpus <- rbind(Trump_after_analysis, Trump_before_analysis, vp_after_analysis, vp_before_analysis,
                biden_after_analysis, biden_before_analysis, kamala_after_analysis, kamala_before_analysis,
                obama_after_analysis, obama_before_analysis, aoc_after_analysis, aoc_before_analysis)
corpus <- corpus %>% mutate(Document=paste(when,"_",screen_name,sep="")) 

###
#creation du corpus
corpus = data.frame(lapply(corpus, as.character), stringsAsFactors=FALSE)

write.csv(corpus, "data supervised_L\\corpus.csv", row.names=FALSE)

```

### clean the corpus
I have seen many things qui vont pas et qu'on doit nettoyer 
```{r}
corpus <- read_csv("data supervised_L/corpus.csv")
corpus <- corpus %>% mutate(Document=paste(when,"_",screen_name,sep="")) 
# Set the text to lowercase
corpus$text <- tolower(corpus$text)
# Remove mentions, urls, emojis, numbers, punctuations, etc.
corpus$text <- gsub("@\\w+", "", corpus$text)
corpus$text <- gsub("\\d+\\w*\\d*", "", corpus$text)
corpus$text <- gsub("#\\w+", "", corpus$text)
corpus$text <- gsub("[^\x01-\x7F]", "", corpus$text)
corpus$text <- gsub("[[:punct:]]", " ", corpus$text)


corpus_final = data.frame(lapply(corpus, as.character), stringsAsFactors=FALSE)

write.csv(corpus_final, "data supervised_L\\corpus_final_before_after.csv", row.names=FALSE)
```

### GLOBAL CORPUS 

```{r}
# trump analysis
TRUMP_all_tweet <- read_csv("data supervised_L/trump_after.csv")

TRUMP_all_tweet <- TRUMP_all_tweet %>% 
  select(text,screen_name) %>% 
  group_by(screen_name) %>% 
  summarise(paste(TRUMP_all_tweet$text, collapse = ' ')) 

TRUMP_all_tweet <- rename(TRUMP_all_tweet, text = `paste(TRUMP_all_tweet$text, collapse = " ")`)

# vp analysis
vp_all_tweet <- read_csv("data supervised_L/vp_after.csv")

vp_all_tweet <- vp_all_tweet %>% 
  select(text,screen_name) %>% 
  group_by(screen_name) %>% 
  summarise(paste(vp_all_tweet$text, collapse = ' ')) 

vp_all_tweet <- rename(vp_all_tweet, text = `paste(vp_all_tweet$text, collapse = " ")`)

# biden analysis
biden_all_tweet <- read_csv("data supervised_L/biden_after.csv")

biden_all_tweet <- biden_all_tweet %>% 
  select(text,screen_name) %>% 
  group_by(screen_name) %>% 
  summarise(paste(biden_all_tweet$text, collapse = ' ')) 

biden_all_tweet <- rename(biden_all_tweet, text = `paste(biden_all_tweet$text, collapse = " ")`)
  
# kamala analysis
kamala_all_tweet <- read_csv("data supervised_L/kamala_after.csv")

kamala_all_tweet <- kamala_all_tweet %>% 
  select(text,screen_name) %>% 
  group_by(screen_name) %>% 
  summarise(paste(kamala_all_tweet$text, collapse = ' ')) 

kamala_all_tweet <- rename(kamala_all_tweet, text = `paste(kamala_all_tweet$text, collapse = " ")`)

# obama analysis
obama_all_tweet <- read_csv("data supervised_L/obama_after.csv")

obama_all_tweet <- obama_all_tweet %>% 
  select(text,screen_name) %>% 
  group_by(screen_name) %>% 
  summarise(paste(obama_all_tweet$text, collapse = ' ')) 

obama_all_tweet <- rename(obama_all_tweet, text = `paste(obama_all_tweet$text, collapse = " ")`)

# AOC analysis
aoc_all_tweet <- read_csv("data supervised_L/aoc_after.csv")

aoc_all_tweet <- aoc_all_tweet %>% 
  select(text,screen_name) %>% 
  group_by(screen_name) %>% 
  summarise(paste(aoc_all_tweet$text, collapse = ' ')) 

aoc_all_tweet <- rename(aoc_all_tweet, text = `paste(aoc_all_tweet$text, collapse = " ")`)

corpus_all <- rbind(TRUMP_all_tweet, vp_all_tweet, biden_all_tweet, kamala_all_tweet, obama_all_tweet, aoc_all_tweet)

###
#creation du corpus
corpus_all = data.frame(lapply(corpus_all, as.character), stringsAsFactors=FALSE)

write.csv(corpus_all, "data supervised_L\\corpus_final.csv", row.names=FALSE)

```

### clean the corpus
I have seen many things qui vont pas et qu'on doit nettoyer 
```{r}
corpus_final <- read_csv("data supervised_L/corpus_final.csv")
# Set the text to lowercase
corpus_final$text <- tolower(corpus_final$text)
# Remove mentions, urls, emojis, numbers, punctuations, etc.
corpus_final$text <- gsub("@\\w+", "", corpus_final$text)
corpus_final$text <- gsub("\\d+\\w*\\d*", "", corpus_final$text)
corpus_final$text <- gsub("#\\w+", "", corpus_final$text)
corpus_final$text <- gsub("[^\x01-\x7F]", "", corpus_final$text)
corpus_final$text <- gsub("[[:punct:]]", " ", corpus_final$text)


corpus_final = data.frame(lapply(corpus_final, as.character), stringsAsFactors=FALSE)

write.csv(corpus_final, "data supervised_L\\corpus_final.csv", row.names=FALSE)
```
